lambdas/pyproject.toml
[tool.poetry]
name = "sapientone-lambdas"
version = "0.1.0"
description = ""
authors = ["William Di Pasquale <wdp@dipasqualew.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "sapientone_lambdas"}]

[tool.poetry.dependencies]
python = "^3.9"
langchain = "^0.0.137"
openai = "^0.27.4"
tiktoken = "^0.3.3"
pinecone-client = "^2.2.1"
pinecone-text = "^0.4.2"
pgvector = "^0.1.6"
psycopg2-binary = "^2.9.6"
pydantic = "^1.10.7"
fastapi = "^0.95.1"
uvicorn = {extras = ["standard"], version = "^0.21.1"}


[tool.poetry.group.dev.dependencies]
ruff = "^0.0.261"
mypy = "^1.2.0"
black = "^23.3.0"
ipykernel = "^6.22.0"
python-dotenv = "^1.0.0"
pytest = "^7.3.0"
pytest-bdd = "^6.1.1"
requests = "^2.28.2"
types-requests = "^2.28.11.17"
httpx = "^0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.ruff]
cache-dir = "../.cache/lambdas/ruff"
line-length = 120

[tool.mypy]
cache_dir = "../.cache/lambdas/mypy"

[[tool.mypy.overrides]]
module = [
    "sqlalchemy",
    "sqlalchemy.orm",
    "pinecone",
]
ignore_missing_imports = true


[tool.pytest.ini_options]
cache_dir = "../.cache/lambdas/pytest"
testpaths = [
    "tests",
]
bdd_features_base_dir = "tests/features/"
markers = [
    "integration: Tests that use real integrations, e.g. OpenAI / Notion",
]
filterwarnings = [
    # Pinecone uses a deprecated function in urllib3, we don't care really
    'ignore:HTTPResponse.getheader():DeprecationWarning:pinecone.core.client.rest',
]
